Added variable height credentialCarousel
Previous version used a PageView element that needs a predetermined height. We do not know the height of the credentialSets before drawing them. So this proposes to use an offstage IndexedStack (which always has the height of the highest element) to obtain the height-value and then creates the PageView with that height.